+++ resources/views/manager/create.blade.php
@@ -4,9 +4,7 @@
 
 <x-chief::page.template :title="$title">
     <x-slot name="hero">
-        <x-chief::page.hero :title="$title" :breadcrumbs="[$resource->getPageBreadCrumb()]" class="max-w-3xl">
-            <button form="createForm" type="submit" class="btn btn-primary">Aanmaken</button>
-        </x-chief::page.hero>
+        <x-chief::page.hero :title="$title" :breadcrumbs="[$resource->getPageBreadCrumb()]" class="max-w-3xl"></x-chief::page.hero>
     </x-slot>
 
     <x-chief::page.grid class="max-w-3xl">
@@ -16,12 +14,15 @@
 
                 <div class="space-y-6">
                     <x-chief-form::fields not-tagged="edit,not-on-create"/>
-
-                    <button type="submit" class="btn btn-primary">Aanmaken</button>
+                    <div class="flex flex-wrap gap-3">
+                        <button type="submit" class="btn btn-primary">Aanmaken</button>
+                    </div>
                 </div>
             </form>
+
         </div>
     </x-chief::page.grid>
+
 </x-chief::page.template>