+++ src/Table/resources/nestable/index.blade.php
@@ -15,15 +15,8 @@
                 {!! $resource->getIndexHeaderContent() !!}
             @endif
 
-            @adminCan('create')
-                <a
-                    href="@adminRoute('create')"
-                    title="{{ ucfirst($resource->getLabel()) }} toevoegen"
-                    class="btn btn-primary"
-                >
-                    <x-chief::icon-label type="add">{{ ucfirst($resource->getLabel()) }} toevoegen</x-chief::icon-label>
-                </a>
-            @endAdminCan
+            @include('chief::manager._index._index_actions')
+
         </x-chief::page.hero>
     </x-slot>